data-collector/requirements.txt
flask==2.2.5
requests==2.31.0
pymongo==4.6.1
grpcio==1.60.0
grpcio-tools==1.60.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
apscheduler==3.10.4
pybreaker==1.0.1
kafka-python==2.0.2
